feat: add reducer  sort events
