+++ src/App.tsx
@@ -1,10 +1,7 @@
 import React, { useEffect } from 'react';
 import { BrowserRouter, Route, Switch } from 'react-router-dom';
 import { connect } from 'react-redux';
-// import { createStore } from 'redux';
-// import reducer from './Redux/reducers';
 import { updateMembers } from './Redux/actions';
-// import AppContext from './context/appContext';
 import MainPage from './Pages/MainPage/MainPage';
 import AddEventPage from './Pages/AddEventPage';
 import Data from './utils/data';
@@ -12,26 +9,11 @@ import { Members } from './interfaces';
 import { MEMBERS, MAIN_URL } from './constants/constants';
 import './App.scss';
 
-// const defaultState = {
-//   isLoad: true,
-//   eventID: '',
-//   currentUser: {} as Members,
-//   users: [] as Members[],
-// };
-
-// const store = createStore(reducer, defaultState);
-
 interface AppProps {
   onFetch: (param: string) => void;
 }
 
 const App: React.FC<AppProps> = ({ onFetch }) => {
-  // const [setMembers] = useState<Members[]>([]);
-
-  // const appContext = {
-  //   members,
-  //   setMembers,
-  // };
 
   useEffect(() => {
     const fetchData = async (): Promise<void> => {
@@ -50,16 +32,12 @@ const App: React.FC<AppProps> = ({ onFetch }) => {
   }, [onFetch]);
 
   return (
-    // <Provider store={store}>
-    // <AppContext.Provider value={appContext}>
     <BrowserRouter>
       <Switch>
         <Route exact path='/' component={MainPage} />
         <Route path='/events' component={AddEventPage} />
       </Switch>
     </BrowserRouter>
-    // </AppContext.Provider>
-    // </Provider>
   );
 };
