+++ src/Pages/MainPage/MainPage.tsx
@@ -1,7 +1,6 @@
-import React, { useState, useEffect } from 'react';
+import React, { useEffect } from 'react';
 import { connect } from 'react-redux';
 import { updateLoading, updateEventID, updateEvents } from '../../Redux/actions';
-import MainContext from '../../context/mainContext';
 import AuthorizationPage from '../AuthorizationPage';
 import HeaderComponent from '../../Components/Header/Header';
 import ControlButtonPanel from '../../Components/ControlButtonPanel/ControlButtonPanel';
@@ -10,7 +9,12 @@ import FooterComponent from '../../Components/Footer/Footer';
 import Data from '../../utils/data';
 import { EVENTS, MAIN_URL } from '../../constants/constants';
 import { Events } from '../../interfaces';
-import { UpdateLoading, UpdateEventID, State, Members } from '../../Redux/interfaces';
+import {
+  UpdateLoading,
+  UpdateEventID,
+  State,
+  Members,
+} from '../../Redux/interfaces';
 
 interface MainPageProps {
   updateLoad: (value: boolean) => UpdateLoading;
@@ -20,12 +24,6 @@ interface MainPageProps {
 }
 
 const MainPage: React.FC<MainPageProps> = ({ updateLoad, updateID, user, onFetch }) => {
-  const [events, setEvents] = useState<Events[]>([]);
-
-  const mainContext = {
-    events,
-    setEvents,
-  };
 
   useEffect(() => {
     const fetchData = async (): Promise<void> => {
@@ -39,7 +37,6 @@ const MainPage: React.FC<MainPageProps> = ({ updateLoad, updateID, user, onFetch
       onFetch(EVENTS);
 
       localStorage.setItem(EVENTS, JSON.stringify(receivedEvents));
-      setEvents(receivedEvents as unknown as Events[]);
       updateLoad(false);
     };
 
@@ -47,18 +44,18 @@ const MainPage: React.FC<MainPageProps> = ({ updateLoad, updateID, user, onFetch
   }, [updateLoad, updateID, onFetch]);
 
   return (
-    <MainContext.Provider value={mainContext}>
-      {user.name
-        ? <>
-          <HeaderComponent />
-          <ControlButtonPanel />
-          <Main />
-          <FooterComponent />
-        </>
-        : <AuthorizationPage />
+    <>
+      {
+        user.name
+          ? <>
+            <HeaderComponent />
+            <ControlButtonPanel />
+            <Main />
+            <FooterComponent />
+          </>
+          : <AuthorizationPage />
       }
-
-    </MainContext.Provider>
+    </>
   );
 };
 
@@ -79,7 +76,6 @@ const mapDispatchToProps = (dispatch: any) => ({
         return res.json();
       })
       .then((json) => {
-        console.log(json);
         const receivedEvents = JSON.parse((json[json.length - 1]).data);
         dispatch(updateEvents(receivedEvents as Events[]))
       });
