--- src/context/mainContext.ts
@@ -1,11 +1,0 @@
-import { createContext } from 'react';
-import { Events } from '../interfaces';
-
-interface Context {
-  events: Events[];
-  setEvents: React.Dispatch<React.SetStateAction<Events[]>>;
-}
-
-const MainContext = createContext({} as Context);
-
-export default MainContext;
