+++ src/index.tsx
@@ -6,18 +6,18 @@ import reducer from './Redux/reducers';
 import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
-import { Members, Events } from './Redux/interfaces';
+// import { Members, Events } from './Redux/interfaces';
 
-const defaultState = {
-  isLoad: true,
-  eventID: '',
-  currentUser: {} as Members,
-  users: [] as Members[],
-  isUpdate: '',
-  events: [] as Events[],
-};
+// const defaultState = {
+//   isLoad: true,
+//   eventID: '',
+//   currentUser: {} as Members,
+//   users: [] as Members[],
+//   isUpdate: '',
+//   events: [] as Events[],
+// };
 
-const store = createStore(reducer, defaultState);
+const store = createStore(reducer);
 
 ReactDOM.render(
   <React.StrictMode>
